functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.0.0
pysam>=0.22.0
google-cloud-storage>=2.0.0
